% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{yq_config}
\alias{yq_config}
\title{Configure YuQue}
\usage{
yq_config()
}
\value{
Nothing.
}
\description{
Ask some questions to user and save configuration to local file.
}
\details{
More details see \link{yq_whoami}, \link{yq_token}.
}
\examples{
\dontrun{
if (interactive())
  yq_config()
}
}
