% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{yq_list_groups}
\alias{yq_list_groups}
\alias{yq_list_public_groups}
\alias{yq_list_group_info}
\title{List Information for Group, Repo etc.}
\usage{
yq_list_groups(verbose = c("less", "all"))

yq_list_public_groups(verbose = c("less", "all"))

yq_list_group_info(login)
}
\arguments{
\item{verbose}{'all' or 'less', the former will print more messages.}

\item{login}{a string for login name of YuQue.}
}
\value{
a \code{list} or \code{data.frame}
}
\description{
List Information for Group, Repo etc.
}
