% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_request.R, R/token.R
\name{yq_token}
\alias{yq_token}
\title{Return the local user's YuQue Personal Access Token (PAT)}
\usage{
yq_token()

yq_token()
}
\value{
A string, with the token, or a zero length string scalar,
if no token is available.

A string, with the token, or a zero length string scalar,
if no token is available.
}
\description{
You can read more about PATs here:
\url{https://www.yuque.com/yuque/developer/api#785a3731}
and you can access your PATs here (if logged in to YuQue):
\url{https://www.yuque.com/settings}.

You can read more about PATs here:
\url{https://www.yuque.com/yuque/developer/api#785a3731}
and you can access your PATs here (if logged in to YuQue):
\url{https://www.yuque.com/settings}.
}
\details{
Currently it consults the \code{YUQUE_TOKEN}
environment variables, in this order.

Currently it consults the \code{YUQUE_TOKEN}
environment variables, in this order.
}
\examples{
yq_token()
}
