% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami.R
\name{yq_whoami}
\alias{yq_whoami}
\title{Info on current YuQue user}
\usage{
yq_whoami(.token = NULL, .api_url = NULL, .send_headers = NULL)
}
\arguments{
\item{.token}{Authentication token. Default to YUQUE_TOKEN
environment variables, in this order if any is set.}

\item{.api_url}{Github API url (default: \url{https://www.yuque.com/api/v2}). Used
if \code{endpoint} just contains a path. Default to YuQue_API_URL
environment variable if set.}

\item{.send_headers}{Named character vector of header field values
(excepting \code{Authorization}, which is handled via
\code{.token}). This can be used to override or augment the
defaults, which are as follows: the \code{Accept} field defaults
to \code{"application/json"} and the
\code{User-Agent} field defaults to
\code{"https://github.com/ShixiangWang/yq"}. This can be used
to, e.g., provide a custom media type, in order to access a
preview feature of the API.}
}
\value{
a \code{data.frame}
}
\description{
Reports basic profile for the YuQue user.
}
\details{
Get a personal access token for the YuQue API from
\url{https://www.yuque.com/settings} and select the scopes necessary for
your planned tasks. The \code{repo} scope, for example, is one many are
likely to need. The token itself is a string of 40 letters and digits.

You can set it using \link{yq_config} function. You can also store it any way
you like and provide explicitly via the \code{.token}
argument to \code{\link{yq}()}.

However, you can also define an environment variable \code{YUQUE_TOKEN}
with this value in their \code{.Renviron} file. Add a
line that looks like this, substituting your PAT:

\preformatted{
YUQUE_TOKEN=8c70fd8419398999c9ac5bacf3192882193cadf2
}

Put a line break at the end! If you're using an editor that shows line
numbers, there should be (at least) two lines, where the second one is empty.
Restart R for this to take effect. Call \code{yq_whoami()} to confirm
success.

To get complete information on the authenticated user, call
\code{yq("/users/your_login_name")}.
}
\examples{
\dontrun{
yq_whoami()

}
}
